Changes the app component interfaces
Changes the app component interfaces to allow for
the integration of other types of data sources,
processing and output. To this end the interfaces
have had their method names made more general and
they have been made generic to accomodate that
need to integrate other types of data and the
processing of that data.