Creates and Implements a DeepCopyable Interface
Creates a DeepCopyable interface  that is
implemented by the WeatherDataPoint class and will
be implemented by future entities if a deep copy
of them can be created. The DeepCopyable interface
requires implementing classes to have a deepCopy
method that created a deep copy of the object on
which the method is called.